decouple quick check status from retrieval
